several color palettes
